FIANLLY MADE IT WORK!!!
I moved the entire module into the plugins folder rather than just the produced jar file, which meant a new mods-mvn folder was created for the module in the plugins folder. The task explains that I should insert only the produced jar file, I think. But I like it better that all foreign modules are in the plugins folder and jars in the plugins\mods-mvn folder for further seperation. The solution was adding "mods-mvn" to the "plugins" in the loader, so it becomes "plugins\\mods-mvn". Good soup!